fix stdin when debugging with lldb and run-in-terminal
Reproducing the problem:

0. Configure Qt Creator to use lldb with run-in-terminal enabled;
1. Create a new console project;
2. Use cin to read some numbers from console, e.g.,

  ```c++
  int a, b;
  cout << "Hello World!" << endl;
  cin >> a >> b;
  cout << a + b << endl;
  ```
3. Start debugging.

Expected:

The program runs in external terminal and functions well.

Reality:

External terminal opens, but the program runs as "run-in-terminal" is
disabled: the program output is shown in "Application Output" where
user input is never received. As a result, the program gets stuck.

The cause:

lldbbridge ignores the `useTerminal_` parameter when running with
the `StartInternal` start mode. It just launches the program (debugee)
by itself rather than attaching to existing `qtcreator_process_stub`
child process. And there is no where a user can give the input to
the debugee.

The fix:

When `startMode_` is `StartInternal` and `useTerminal_` is set, just
reuse the `AttachExternal` logic to attach the `qtcreator_process_stub`
child process, which is ready to run the debugee. Reattaching seems to
be benign here.

Known issue:

There seems always be a breakpoint at main().